renaming babel-plugin-card-file-analyze to analyze
